+++ include/cufinufft/common.h
