+++ test/spread1d_test.cu
@@ -4,7 +4,7 @@
 #include <helper_cuda.h>
 #include <complex>
 #include <algorithm>
-#include "../src/cuspreadinterp.h"
+#include "cufinufft/cuspreadinterp.h"
 #include "../contrib/utils.h"
 
 using namespace std;
